LayoutTests/imported/w3c/web-platform-tests/uievents/mouse/mouse_boundary_events_after_reappending_last_over_target.tentative-expected.txt


PASS After re-appending the last over element at mouseover, mouse boundary events should be fired as just over on the target
PASS After re-appending the last over element at mouseenter, mouse boundary events should not be fired
PASS After re-appending the last over element after mouseover, mouse boundary events should not be fired
FAIL After re-appending the last over element to outside the deepest mouseenter target (but keeps it as under the cursor) at mouseover, mouse boundary events should be fired on it again to correct the following mouseleave targets assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseleave@div#parent, mouseover@div#child, mouseenter@div#child" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
FAIL After re-appending the last over element to outside the deepest mouseenter target at mouseover, mouse boundary events should be fired only on the element under the cursor assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
PASS After re-appending the last over element to the deepest mouseenter target without the original parent at mouseover, mouse boundary events should not be fired
FAIL After re-appending the last over element to outside the deepest mouseenter target (but keeps under the cursor) after mouseover, mouse boundary events should be fired on it again to correct the following mouseleave event targets assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseleave@div#parent, mouseover@div#child, mouseenter@div#child" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
FAIL After re-appending the last over element to outside the deepest mouseenter target after mouseover, mouse boundary events should be fired only on the element under the cursor assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
PASS After re-appending the last over element to the deepest mouseenter target without the original parent after mouseover, mouse boundary events should not be fired
PASS After removing and re-appending the last over element with flushing layout after mouseover, mouse boundary events should not be fired
FAIL After removing after mouseover and re-appending the last over element at next animation frame, mouse boundary events should be fired assert_equals: expected "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child, mouseover@div#parent, mouseout@div#parent, mouseover@div#child, mouseenter@div#child" but got "mouseover@div#child, mouseenter@body, mouseenter@div#grandparent, mouseenter@div#parent, mouseenter@div#child, mousemove@div#child"
